Implement TodoApplication (connected with sqlite3/postgreSQL), add endpoints on adding users and todos and accessing information as an admin

+++ 02_fastapi/TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from typing import Annotated
 from database import sessionLocal
 from pydantic import BaseModel, Field
 from routers import auth
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -28,28 +30,38 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    return db.query(Todos).filter(Todos.owner == user.get("id")).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get('id')).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status_code=404, detail="Todo not found!")
+    raise HTTPException(status_code=404, detail="Todo not found!") 
 
 
 @router.post("/todo/", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    todo_model = Todos(**todo_request.model_dump(), owner= user.get('id'))
     db.add(todo_model)
     db.commit()
+    db.refresh(todo_model)
+    return todo_model
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0),):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found, try a different input.")
     todo_model.title = todo_request.title
@@ -62,9 +74,11 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed.")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found. Try with a different input!")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner == user.get("id")).delete()
     db.commit()
